% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aufgaben.R
\name{generate_numbers}
\alias{generate_numbers}
\title{Generate random values from a normal distribution}
\usage{
generate_numbers(n, mean, sd, min = -Inf, max = Inf, precision = 2, try = 1)
}
\arguments{
\item{n}{The amount of values to be generated}

\item{mean}{The mean of the normal distribution to be sampled from}

\item{sd}{The standard deviation of the normal distribution to be sampled from}

\item{min}{The smallest acceptable value (defaults to -Inf)}

\item{max}{The largest acceptable value (defaults to Inf)}

\item{precision}{Decimals to be rounded to (defaults to 2)}
}
\value{
Some random values
}
\description{
Generate random values from a normal distribution
}
