% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable.R
\name{get_crosstable}
\alias{get_crosstable}
\title{Prepare crosstable from a dataset}
\usage{
get_crosstable(data, row_order = NULL, col_order = NULL)
}
\arguments{
\item{data}{Long-form dataset (tibble) with one or two nominal type variables}

\item{row_order}{Optional vector to specify the order of the first variable}

\item{col_order}{Optional vector to specify the order of the second variable}
}
\value{
A list containing a specially formatted crosstable and additional information for use in \code{print_crosstable()}
}
\description{
Prepare crosstable from a dataset
}
